+++ src/lib.rs
@@ -179,6 +179,7 @@ pub mod dtm;
 pub mod targets;
 pub mod hdl;
 pub mod writable;
+pub mod apps;
 pub mod ffi;
 pub mod cim;
 pub mod oim;